Move extract methods to ExportFile
