+++ GUI/Types/Exporter/ExportFile.cs
@@ -3,7 +3,9 @@ using System.IO;
 using System.Windows.Forms;
 using GUI.Forms;
 using GUI.Utils;
+using SteamDatabase.ValvePak;
 using ValveResourceFormat.IO;
+using Resource = ValveResourceFormat.Resource;
 
 namespace GUI.Types.Exporter
 {
@@ -30,7 +32,7 @@ namespace GUI.Types.Exporter
                 filter = $"{gltfFilter}|{glbFilter}|{filter}";
             }
 
-            var dialog = new SaveFileDialog
+            using var dialog = new SaveFileDialog
             {
                 FileName = Path.GetFileNameWithoutExtension(fileName),
                 InitialDirectory = Settings.Config.SaveDirectory,
@@ -82,6 +84,83 @@ namespace GUI.Types.Exporter
                 Console.WriteLine($"Export for \"{fileName}\" completed");
             };
             extractDialog.ShowDialog();
+        }
+
+        public static void ExtractFileFromPackageEntry(PackageEntry file, VrfGuiContext vrfGuiContext, bool decompile)
+        {
+            vrfGuiContext.CurrentPackage.ReadEntry(file, out var output, validateCrc: file.CRC32 > 0);
+
+            ExtractFileFromByteArray(file.GetFileName(), output, vrfGuiContext, decompile);
+        }
+
+        public static void ExtractFileFromByteArray(string fileName, byte[] output, VrfGuiContext vrfGuiContext, bool decompile)
+        {
+            if (decompile && fileName.EndsWith("_c", StringComparison.Ordinal))
+            {
+                using var resource = new Resource
+                {
+                    FileName = fileName,
+                };
+                using var memory = new MemoryStream(output);
+
+                resource.Read(memory);
+
+                Export(fileName, new ExportData
+                {
+                    Resource = resource,
+                    VrfGuiContext = new VrfGuiContext(null, vrfGuiContext),
+                });
+
+                return;
+            }
+
+            var dialog = new SaveFileDialog
+            {
+                InitialDirectory = Settings.Config.SaveDirectory,
+                Filter = "All files (*.*)|*.*",
+                FileName = fileName,
+            };
+            var userOK = dialog.ShowDialog();
+
+            if (userOK == DialogResult.OK)
+            {
+                Settings.Config.SaveDirectory = Path.GetDirectoryName(dialog.FileName);
+                Settings.Save();
+
+                using var stream = dialog.OpenFile();
+                stream.Write(output, 0, output.Length);
+            }
+        }
+
+        public static void ExtractFilesFromTreeNode(TreeNode selectedNode, VrfGuiContext vrfGuiContext, bool decompile)
+        {
+            if (selectedNode.Tag is PackageEntry file)
+            {
+                // We are a file
+                ExtractFileFromPackageEntry(file, vrfGuiContext, decompile);
+            }
+            else
+            {
+                // We are a folder
+                using var dialog = new FolderBrowserDialog();
+
+                if (dialog.ShowDialog() != DialogResult.OK)
+                {
+                    return;
+                }
+
+                ExportData exportData = null;
+                if (decompile)
+                {
+                    exportData = new ExportData
+                    {
+                        VrfGuiContext = new VrfGuiContext(null, vrfGuiContext),
+                    };
+                }
+
+                var extractDialog = new ExtractProgressForm(vrfGuiContext.CurrentPackage, selectedNode, dialog.SelectedPath, exportData);
+                extractDialog.ShowDialog();
+            }
         }
     }
 }
